Refactored files' path regarding CBS
